change models name to singular, set up error handling middleware, remove req.user line, rename middleware to isAdmin
