+++ leaderboard/controllers/weeks.js
@@ -1,4 +1,4 @@
-const Week = require('../models/weeks');
+const Week = require('../models/week');
 module.exports = {
 
 	/* GET /admin/weeks/new */
@@ -16,36 +16,26 @@ module.exports = {
 	 			{ place: '3rd', name: req.body.third }
  			]
  		};
- 		await Week.create(newWeek, (err, createdWeek) => {
- 			if (err) {
- 				console.log(err);
- 			} else {
- 				console.log(createdWeek);
-				res.redirect('/admin/weeks');
- 			}
- 		});	
+ 		const week = await Week.create(newWeek);
+		console.log(week);
+		res.redirect('/admin/weeks'); 			
  	},
 
 	/* GET /admin/weeks */
 	async getAdminWeeks (req, res, next) {
-		await Week.find({}, (err, weeks) => {
-			if (err) {
-				console.log(err);
-			} else {
-				res.render('weeks/weeks', { weeks });
-			}
-		});
+		const weeks = await Week.find({});
+		res.render('weeks/weeks', { weeks });
 	},
 
 	/* GET admin/weeks/:id/edit */
 	async getEditWeek (req, res, next) {
-		let week = await Week.findById(req.params.id);
+		const week = await Week.findById(req.params.id);
 	  res.render('weeks/edit', { week });
 	},
 
 	/* PUT admin/weeks/:id */
 	async updateWeek (req, res, next) {
-	  let week = await Week.findById(req.params.id);
+	  const week = await Week.findById(req.params.id);
 	  week.winners = [
 	 			{ place: '1st', name: req.body.first },
 	 			{ place: '2nd', name: req.body.second },
@@ -58,12 +48,7 @@ module.exports = {
 
 	/* DELETE admin/weeks/:id */
 	async deleteWeek (req, res, next) {
-		await Week.findByIdAndRemove(req.params.id, (err) => {
-			if (err) {
-				console.log(err);
-			} else {
-				res.redirect('back');				
-			}
-		});
+		await Week.findByIdAndRemove(req.params.id);
+		res.redirect('back');
 	}
 };
