+++ leaderboard/routes/weeks.js
@@ -2,7 +2,7 @@ var express = require('express');
 var router = express.Router();
 var { asyncErrorHandler,
 			isLoggedIn,
-			requireAdmin } = require('../middleware');
+			isAdmin } = require('../middleware');
 var {
 	postAdminWeek,
 	getAdminWeeks,
@@ -13,21 +13,21 @@ var {
 } = require('../controllers/weeks');
 
 /* GET /admin/weeks/new */
-router.get('/new', isLoggedIn, requireAdmin, getNewWeek);
+router.get('/new', isLoggedIn, isAdmin, getNewWeek);
 
 /* POST /admin/weeks */
-router.post('/', isLoggedIn, requireAdmin, asyncErrorHandler(postAdminWeek));
+router.post('/', isLoggedIn, isAdmin, asyncErrorHandler(postAdminWeek));
 
 /* GET /admin/weeks */
-router.get('/', isLoggedIn, requireAdmin, asyncErrorHandler(getAdminWeeks));
+router.get('/', isLoggedIn, isAdmin, asyncErrorHandler(getAdminWeeks));
 
 /* GET admin/weeks/:id/edit */
-router.get('/:id/edit', isLoggedIn, requireAdmin, asyncErrorHandler(getEditWeek));
+router.get('/:id/edit', isLoggedIn, isAdmin, asyncErrorHandler(getEditWeek));
 
 /* PUT admin/weeks/:id */
-router.put('/:id', isLoggedIn, requireAdmin, asyncErrorHandler(updateWeek));
+router.put('/:id', isLoggedIn, isAdmin, asyncErrorHandler(updateWeek));
 
 /* DELETE admin/weeks/:id */
-router.delete('/:id', isLoggedIn, requireAdmin, asyncErrorHandler(deleteWeek));
+router.delete('/:id', isLoggedIn, isAdmin, asyncErrorHandler(deleteWeek));
 
 module.exports = router;
